microservices/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
websockets==14.2
pydantic_settings==2.8.1
azure-eventgrid==4.11.0
azure-communication-callautomation==1.4.0b1
azure-communication-jobrouter==1.0.0
azure-core==1.32.0
azure-cosmos==4.9.0
